chore: organized user profile components and routes
