+++ src/pages/me/Profile.tsx
@@ -1,0 +1,16 @@
+/**
+ * Profile Page
+ */
+
+// Dependencies
+import WorkInProgress from '../../components/reusable/WorkInProgress';
+
+function Profile() {
+	return (
+		<div>
+			<WorkInProgress />
+		</div>
+	);
+}
+
+export default Profile;
