+++ src/pages/me/ScoreHistory.tsx
@@ -1,0 +1,16 @@
+/**
+ * Score History Page
+ */
+
+// Dependencies
+import WorkInProgress from '../../components/reusable/WorkInProgress';
+
+function ScoreHistory() {
+	return (
+		<div>
+			<WorkInProgress />
+		</div>
+	);
+}
+
+export default ScoreHistory;
